% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSlurmDuration.R
\name{parseSlurmDuration}
\alias{parseSlurmDuration}
\title{Parse a SLURM accounting time duration to an ISO 8601 time duration.}
\usage{
parseSlurmDuration(duration)
}
\arguments{
\item{duration}{a character vector with one element.}
}
\value{
a character vector with one element.
}
\description{
Parse a SLURM accounting time duration to an ISO 8601 time duration.
}
\examples{
parseSlurmDuration("23720-4:20:69")
}
